# Use the most space-efficient serialization format in dense stores