When serializing a dense store with the custom encoding, figure out if the data would be more space-efficiently encoded densely (`BinEncodingContiguousCounts`) or sparsely (`BinEncodingIndexDeltasAndCounts`), and use the better option.

There is a computational cost, but I think it is small enough compared to the space efficiency gain to justify this change.

# Benchmarks

Outputs of `BenchmarkEncode` and `TestBenchmarkEncodedSize` respectively.

## Before

```
BenchmarkEncode/dense/empty/custom-12                                         	19677914	        53.90 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/dense/small_int_count/custom-12                               	 2305779	       540.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/dense/small_non_int_count/custom-12                           	 2044314	       575.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/dense/small_far_apart/custom-12                             	  144552	      8058 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/collapsing_lowest_dense/log_normal_non_int_count/custom-12  	  103376	     11605 ns/op	       0 B/op	       0 allocs/op

ddsketch_test.go:672: test case                                         proto custom custom_no_mapping
ddsketch_test.go:680: dense/empty                                          17     17                 0
ddsketch_test.go:680: dense/small_int_count                              1116    163               146
ddsketch_test.go:680: dense/small_non_int_count                          1116    194               177
ddsketch_test.go:680: dense/small_far_apart                             18631   2364              2347
ddsketch_test.go:680: collapsing_lowest_dense/log_normal_non_int_count   6912   6325              6308
```

## After

```
BenchmarkEncode/dense/empty/custom-12                                       	19616906	        59.18 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/dense/small_int_count/custom-12                                	 1969686	       599.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/dense/small_non_int_count/custom-12                          	 1925890	       632.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/dense/small_far_apart/custom-12                             	  143432	      8171 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncode/collapsing_lowest_dense/log_normal_non_int_count/custom-12  	   85789	     14218 ns/op	       0 B/op	       0 allocs/op

ddsketch_test.go:672: test case                                         proto custom custom_no_mapping
ddsketch_test.go:680: dense/empty                                          17     17                 0
ddsketch_test.go:680: dense/small_int_count                              1116     30                13
ddsketch_test.go:680: dense/small_non_int_count                          1116     61                44
ddsketch_test.go:680: dense/small_far_apart                             18631     40                23
ddsketch_test.go:680: collapsing_lowest_dense/log_normal_non_int_count   6912   6325              6308
```